requirements.txt
backports.csv==1.0.7
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.3
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==0.0.8
django-cleanup==3.2.0
django-crispy-forms==1.7.2
django-filter==2.1.0
django-import-export==1.2.0
django-tables2==2.0.6
et-xmlfile==1.0.1
jdcal==1.4.1
mysqlclient==1.4.2
odfpy==1.4.0
openpyxl==2.6.2
pytz==2019.1
PyYAML==5.1.1
sqlparse==0.3.0
tablib==0.13.0
xlrd==1.2.0
xlwt==1.3.0
